codegen: Handle delegate_target attribute of fields
Delegate fields without a delegate target don't require special handling
on copy/destroy.

Fixes https://gitlab.gnome.org/GNOME/vala/issues/520

+++ codegen/valaccodeassignmentmodule.vala
@@ -221,7 +221,7 @@ public class Vala.CCodeAssignmentModule : CCodeMemberAccessModule {
 		if (lvalue.actual_value_type != null) {
 			type = lvalue.actual_value_type;
 		}
-		if (requires_destroy (type)) {
+		if (get_ccode_delegate_target (field) && requires_destroy (type)) {
 			/* unref old value */
 			ccode.add_expression (destroy_field (field, instance));
 		}
